force the use of QDBusTrayIcon
QDBusTrayIcon implements DBUS Desktop Notifications, and the
XEmbed-based systray implementation that is used as a fallback, uses
"balloons". Ideally, it should be possible to use the DBUS notifications
while still using the DBUS notifications, as many standalone tray
implementations are XEmbed-based.

However, I don't use a system tray, and it appears the main reason I
wanted to scratch this itch,
https://github.com/qutebrowser/qutebrowser/issues/4780 , can and will be
fixed in a better way ... I should have checked that issue before I spent
my night hacking in Qt source code ... Anyway, it still solves the issue
in the meantime.